(PUP-1072) move proxy helper code to util
